requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
mysqlclient==2.2.7
pillow==11.2.1
propcache==0.3.2
PyJWT==2.9.0
python-decouple==3.8
requests==2.32.4
sqlparse==0.5.3
twilio==9.6.3
urllib3==2.5.0
yarl==1.20.1
